Redux test environment setup and test fix
