+++ webapp/src/__tests__/AboutPage.test.tsx
@@ -2,6 +2,7 @@ import About from "../components/About";
 import { screen  } from "@testing-library/react"
 import {BrowserRouter} from "react-router-dom";
 import { render } from '../setupTests'
+import '@testing-library/jest-dom'
 
 
 /**
